move one exception class
